requirements.txt
python-pptx
pillow
#googletrans==4.0.0rc1
#googletrans==4.0.2
#googletrans==3.1.0a0
numpy
xmltodict
